[Core] Moved checks.py to main core package
